Fix a test to not try to run gradle
SetSdkDirHyperlinkTest implicitly runs Gradle Sync, through the use of
`executeIfClicked` which triggers a sync to pick up the changed state
of the project build files.  However, SetSdkDirHyperlinkTest is in the
_other split of the Gradle project system tests, which does not have
the gradle software accessible.  The test passes because it does not
in fact assert anything about the result of Gradle sync, or even that
it started successfully, so make that explicit by replacing
GradleSyncInvoker with its fake counterpart.

Bug: 325936093
Test: existing
Change-Id: Ia570667562ed814e9fa223e0f9b662c29a35d100

GitOrigin-RevId: fdb8ef07b82d53fac1324dee3ad6a2d061d02d94