Only show asterisk when there are changes that can be
... applied by the user.

When an exercise is ongoing, the user should not be able to change the
availability of the sensors. This CL changes the behaviour to fallback
onto the device's most up-to-date state to show the availability.
Furthermore, we only show the asterisk when there is a pending override
change.

The CL also ensures that the asterisk is not shown if there is a
pending override change once there is no longer an ongoing exercise.
This is because the user cannot apply those changes.

These changes aim at reducing confusion over the use of the asterisk.

Fixes: 338446522
Test: //tools/adt/idea/wear-whs:intellij.android.wear-whs_tests
Change-Id: I9b6d02e8157a5683ecddad689e35b38d49c5e66a

GitOrigin-RevId: 13a35fd57326b8ab8bd4aefe5da3f7a0b1ea7b59