Adjust Blueprint mode text sizes
The font size calculation is just an approximation and, at the moment,
was way off. The original code intended for the blueprint text to be
similar size to the rendered size but this is wrong since Swing uses
point font sizes and not pixels.

This change reduces the point sizes to make it work better.

Fixes: 364539542
Test: N/A
Change-Id: I80b22c8f2b9ca103c7fc309d139bd0c8942e2104

GitOrigin-RevId: 6b87ec862f61fc0dd95f76461e0bfb22a3422c04