[android] use LF line separators instead of CRLF in testData files
Currently, line separators in text files are specified via git config. This isn't convenient because it requires additional configuration on each location where the source is checked out. It would be more convenient to specify line separators via .gitattributes file, but for that we need to ensure that line separators in files are consistent with .gitattributes. Most of the files already use LF separators, this commit converts line separator to LF for several files which use CRLF.

GitOrigin-RevId: 7ad36716c8f1baa07699c20439cba051170589d5