[android] specify line separators for native-symbolizer/testData/win/symbols.txt via .gitattributes
Currently, line separators in text files are specified via git config. This isn't convenient because it requires additional configuration on each location where the source is checked out. It would be more convenient to specify line separators via .gitattributes file, but for that we need to ensure that line separators in files are consistent with .gitattributes. We're going to set LF separators by default, so CRLF separators for bin/win/symbols.txt should be specified explicitly.

GitOrigin-RevId: a09b70b57e0679a992138f1fbf885af6b7ebc51e