[android] make line separators in native/installer/win consistent
Currently, line separators in text files are specified via git config. This isn't convenient because it requires additional configuration on each location where the source is checked out. It would be more convenient to specify line separators via .gitattributes file, but for that we need to ensure that line separators in files are consistent with .gitattributes. We're going to set LF separators by default, so CRLF separators for README.txt files are now set explicitly in .gitattributes, and UAC.nsh is converted to use LF separators like other *.nsh files.

GitOrigin-RevId: dd666b982636efa8e1c566cda2ba4b62f7d5bf3e