Check if changed file belongs to the current project
In PsiCodeFileChangeDetectorService, we were incorrectly detecting
changes to files from a different project. The reason was we were
relying on the assumption that PsiFile#getProject() would return
the project that contains the file. However, this assumption is false
and the method will return whichever Project is set to the current
PsiManager.

To fix this issue, we need to filter out files from different projects
when detecting code changes using the PsiManager#isInProject() method.

Bug: 361721942
Test: PsiCodeFileChangeDetectorServiceTest
Change-Id: Id224cfc030021deaa857a4a87e4cb6743650f64f

GitOrigin-RevId: d64a049074801d69b5b2198b56d4c205dbb43e9e