Add 'Ask Gemini' link on MessageEvents with level ERROR or above
To add the link, previous implementation depended on the BuildEvent
object's parentId to be an instance of ExternalSystemTaskId and deriving
the task type from this instance. However, this only works for top level
messages. Many errors are not on the top level but related to a
task, these have task name (String) as the parent ID.

However,as parentId can also be a
String, this causes the link to not be added, and resulting in bugs like
b/330441116, b/295887979.

We are now relying on the message's severity being ERROR in-order to add
the Ask Gemini link.

bug: 330441116, b/295887979
fixes: 330441116, b/295887979
Test: BuildOutputParserWrapperTest
Change-Id: I8aa54e5bcc9019c981195d2a5aa4ddf7594da77e

GitOrigin-RevId: d466e96e9d520b9308bcbdee3c9462ea2c0d45f8