Add SimpleFileLinkFilter as Last Filter
SimpleFileLinkFilter is a intended to catch logs that emit `File:###`
explicitly to trigger a link. For example, Timbr has a way to log
with a TAG like this.

This pattern will also match lines from exceptions, but will not
include the package name so we get an ambiguous link.

Moving this filter from first to last will ensure the exception filters
will match and short-circuit.

Fixes: 365049153
Test: Updated
Change-Id: Ie35dcc6a078f56a8c939ad6efa1208caa734ad07

GitOrigin-RevId: e40eb2bde8d4ba109c4348f986984533c263f19b