Make RoomSchemaManager use a more generic scope concept
Use the test-or-otherwise status of the file being queried to know
whether to look up the room schema in a test-enabled-or-not cache.

This is not exactly a one-to-one behaviour-neutral change.
Previously, each module would have a RoomSchemaManager with 5 entries,
corresponding to one for each of the possible
AndroidModuleSystem.ScopeType values; now, each module has a
RoomSchemaManager with 2 entries, corresponding to whether the scope
should include tests or not for this module.  For a
module-per-sourceset system, we have moved from 4 entries
which *should* never be looked up to 1 (as each sourceset has either
test or production sources, not both).

Bug: none filed
Test: existing, I hope
Change-Id: I35fa1a98d55a642e7e97fc611606ff0449f55770

GitOrigin-RevId: 79cff50a581be4e87b149a3282a11be482330470