[android] IDEA-358976 register default NoOpStudioBot application service in IntelliJ IDEA environment
Android plugin's StudioBot implementation doesn't differentiate Android Studio from IntelliJ environment. By default, AS considers that the execution context is always in the Android Studio; hence StudioBot is always invoked by default.

In IJ IDEA environment, we have no access to StudioBot APIs and don't want to utilize them. Because of this, the default NoOpStudioBot service is registered in IntelliJ environment only.

`NoOpStudioBot` always considers StudioBot as disabled, and all project files are 'added' to the StudioBot exclude list.

In addition to this, GeminiEditorActionGroup registration is moved to Android Studio configuration files, so that GeminiEditorActionGroup doesn't get registered in IntelliJ.

GitOrigin-RevId: a4f372fccf5737cdec1b26311ce81445c8ebd354