Use `R` reference for resolving resource
When doing code folding for `R.string.blah`, we're currently always
using resources from the current module. That logic dates from when
transitive resources were in use, and so it always worked fine.

Now with non-transitive resources, this doesn't work when the resource
is coming from an R class in a different module/library/whatever. By
getting the repository for whatever module the reference points to, we
ensure the value can be resolved.

I expect this to fix b/356171302 indirectly, by removing the unresolved
folded regions that were causing problems.

Fixes: 356171302
Test: AndroidResourceFoldingTestGenerated, manual verification
Change-Id: Ic4ae450522cf3f7f9e177aeb6772c4d8c6aaed32

GitOrigin-RevId: edb63b6a10ab4c2825cf8da9dc14087288d9b4fc