Hide some uses of getHolderModule()
Several parts of the IDE wish to get a name for the
conceptually-linked (and, in Gradle, actually-linked) set of modules
made up of several associated sourceSets.  That's naturally expressed
as the name of the holder module, except that the generic IDE should
not need to know about the implementation detail of holder modules, so
provide a module system method that does the appropriate thing by
default and in Gradle projects.

Bug: 147586218
Test: existing
Change-Id: I9849f13c453f24c6e70581b450a0dd6ffff3c012

GitOrigin-RevId: 3060c9166f7c74469255a484c8b07148eef7f0ca