Avoid traversing module graph on EDT
AndroidPsiTreeChangeListener is used to aggregate a lot of handling
around PSI changes, and it implements PsiTreeChangeListener. But
PsiTreeChangeListener explicitly says not to us it "at all cost",
instead preferring CachedValues.

For the specific case of AttributeDefinitions, the transition is
straightforward. The definitions are computed based upon a resource
manager, which already exposes a modification count that is updated
whenever the attr resources change. (And in fact, it's smarter than the
existing logic and doesn't update if no resources have changed - so this
may be a net reduction in the number of invalidations.)

The repository's modification count implicitly takes care of changes
that happen both in the current module and any dependent modules, so the
logic to traverse modules that used to exist in
AndroidPsiTreeChangeListener is no longer necessary.

There's still an explicit invalidation of the definitions that happens
when the first build completes, which I think is still necessary. That
can be more easily done now, though, by simply replacing the cached
value.

Fixes: 361784551
Test: Covered by AttributeDefinitionsTest and other existing tests
Change-Id: I31dd648c0c1b1de29a33b4223feb4d1e867daeb7

GitOrigin-RevId: fbf1a0efcef9fe82b681d176c6136c1a73eb4ba4