Change convention in AndroidRunConfiguration
I don't think it matters any more which module of a module group the
AndroidRunConfiguration is created with: methods within the
configuration implementation call into the project system to retrieve
the appropriate information.  Given that, use the holder modules
instead, because at least they display relatively nicely in UI.

However, this means that we need to inform the
ModuleBasedConfiguration more specifically which modules should be
built in the MakeBeforeRunTask.  Define a getModules() method which
converts to main modules if the profile is not a test configuration,
or holder modules (explicitly for now, though it could be implicit) if
it is.

Provide a migration in readExternal to convert previously-externalised
run configurations from the sourceSet module to the holder module
representation.

Bug: 147586218
Test: snapshots and other tests adjusted
Test: manually verified Compose Preview
Change-Id: Ifdce862d7c1ece0d974c900dc7aae7eb15997aad

GitOrigin-RevId: 03f2aff16ba4430e866894310b4da77555afc075