Add faster access to Show Bytecode in Apk/DexFileViewer
This commit adds double-click on leaf nodes to quickly see the Dex
Bytecode in the DexFileViewer of the ApkViewer. It also adds the
space key as the keyboard fast access to the selected node, if it's
a valid Dex node.

Note the space key is not configurable, since it's a context/focus-
sensitive action, and there are very limited actions in the
DexViewer.

Fixes: 79522585
Test: manual
Change-Id: I2f218917cb248dd61febb91b6aba817ac7b164a8

GitOrigin-RevId: 3b1eb12e987e9d9e92f1204c7f846762e8bce140