Propagate isTestConfiguration through constructors
Overriding the isTestConfiguration() method is not sound, because as
part of construction of AndroidRunConfigurationBase we take the
constructor parameter and use it to initialize the
AndroidRunConfigurationModule, which therefore ends up inconsistent
with the ComposePreviewRunConfiguration.  This may not have any
immediately obvious implications (there might be some slight changes
in behaviour related to re-running failed tests or other similar
platform facilities), but we can address this by defining a secondary
constructor and threading the initialization parameter through.

Enforce this by making the isTestConfiguration() method final.

Bug: 147586218
Test: existing
Change-Id: Ifd64561028a150057a7e28b37ed9dcc091ceae7b

GitOrigin-RevId: 75127c8be5f589461f4a740b4a0cda1150b03ae9