Ask users to retry the AEHD installer if users fail to grant the administrator privilege to the installer.
If users accidentally click "no" when the AEHD installer prompt
for administrator privilege or don't even notice the prompt before
it times out, the ShellExecuteEx API will return ERROR_CANCELLED
(1223). In this case, ask users to retry the installer in the
installer's error message.
ALso log the command line and error code from ShellExecuteEx in
idea.log for debugging purpose.

Bug: 293138141
Change-Id: I3349410232967ac69b5c9e0014f9b80832f07b55

GitOrigin-RevId: 192103b2cc2cd8587495c2a0fa86873bf84f50d0