Fix using wrong fallback value for locale
This was a copy/paste mistake from I47ba9eba07682ba0e

It didn't cause any problem because annotation
findAttributeValue("locale")?.evaluateString() was returning an empty
string so we didn't end up using the default device value for the
locale.

Furthermore, we didn't set a fallback/default locale value in the
Preview annotation, so null was an acceptable fallback value. This CL
fixes the fallback value in case we ever decide to set one in the
Preview annotation.

This CL also replaces the attribute names with constants.

Bug: None
Test: existing
Change-Id: I3e8d3827168a09035cb3603b47a9bdb53b9284ca

GitOrigin-RevId: 10fe795c6e3b6d5f98234db602614367d0ecbeb6