Simplify a dubious utility predicate
AndroidUtils.equal is called in exactly two places in our codebase:
each time with non-nullable Strings, and each time with
distiguishDelimiters as false.

Since the semantics of the predicate are dubious (surely ":a;b" and
"c,d!" should not be considered equal under basically any
circumstances?), tighten up these requirements in the hope that no-one
else ever calls this.

Bug: none filed
Test: existing
Change-Id: I8f5d59e730fea5ecd407e66262e82fcd69a21801

GitOrigin-RevId: d42438ddfcd00c94a2fdb5af12ecdfd1224d0250