Use quickfixes for 'Ask Gemini' links
Previous implementation involved inserting 'Ask Gemini' text
within the error message and then finding this text and
converting  it to link separately in a filter. With this approach, it
is not  possible to send more context in the query.

This change uses BuildIssueQuickFixes instead to represent the
links. The BuildEvents hold more context that we can provide in
the chat query in future.

Bug: 316057751
Test: BuildOutputParsersIntegrationTest.java
Change-Id: I481af83d1b83246b3b0aeb723a4929612cd7b9f4

GitOrigin-RevId: abf678725c5b887f8f9345d2f7d869068bde3474