Avoid fetch insight on variant selection
This change partly reverts a previous change that fetched insight after
the ListEvents call. According to the firebase team, Studio should avoid
calling Titan on variant selection because 1) the variant field doesn't
do anything right now, and 2) save TPU cost. And since the eventId field
doesn't appear to be that important either - switching events on the
console doesn't trigger a new insight - the simplest way to accomplish
this is to revert to the way insights were fetched before.

Test: updated tests.
Fixes: 366571456
Change-Id: I8f883348b87e02900749fc3e5bebf6ad6c104232

GitOrigin-RevId: a49e46dff8c8d09e2573279e085c93b79021773c