Cache insights
With this change, insights are cached as they are loaded, and the cache
is available until the end of the Studio session.

Clicking on refresh button will force re-fetch (and re-caching) of the
insight. For users who manually opt for an enhanced insight, it will be
cached and shown for all future selections of the same issue, avoiding
the need to show the non-enhanced version of the insight and its
accompanying UX.

Test: added new tests.
Fixes: 366571808
Change-Id: I610da8d0bf7db5db1ce482d1b94abb1dfc9cfd08

GitOrigin-RevId: a704b999accc39adef43c8c7c8cefcca69eabf19