Use `R` class as context for goto resolution
ag/24300777 updated resource "go to declaration" logic to use the `R`
class as resolution context in androidTest modules (which correctly
points to the defining module for the `R` class and its resources)
instead of the source element (which points to the current module, which
may not directly contain the resources). That fix was targeted to
androidTest modules, but can equally be applied to all modules - and in
fact, it seems more correct to do it for all modules.

I stayed away from making the larger change before because it wasn't
necessary, and I've learned not to modify the resource subsystem unless
it's needed for a specific reason. But since we're now trying to
abstract away the notion of androidTest outside of Gradle code, it makes
sense to have this change in place for all modules.

Bug: 294431294
Bug: 147586218
Test: TestRClassesTest
Change-Id: Ie4f5f9cae93ed5027de68ce66c694e15f1641cca

GitOrigin-RevId: c91e8b34a92532e2690ba6608cd3c57c5dc12e37