Move KotlinMultiplatformAndroidTestConfigurationProducer
Although it's not explicitly a Gradle-specific producer, in fact
AndroidTestRunConfiguration is a module- and facet-based
configuration, which in practice is tangled up with the
Gradle-specific representation of modules and sourcesets; despite the
name, it's not generic, and other project systems implement their own
run configuration producers.

Moving the producer to the Gradle Project System module doesn't (at
present, anyway) hide it from non-Gradle projects: it merely means
that the code written to implement it can use Gradle-specific methods.

Bug: 147586218
Test: existing
Change-Id: I3f7d02f52573cae3b407c2f26040421624274b78

GitOrigin-RevId: ac74f9778871e0667aca082d398f35d7bf57580b