Apply the project system token hammer to the facet finder nail
When trying to find a suitable project entity (facet, module, or
something of that ilk) for a global process for which we do not have a
known applicationId, we examine source providers of android facets for
their manifests to find candidates that might have launched a process
of that name.

We then need to map that manifest to a module.  For Gradle projects,
we need to provide the sourceSet modules, because the result will be
more accurate if we know whether the thing that we have is known to be
directly deployable or not, and whether it has test nature or not.
Non-Gradle projects at present have a best-effort equivalent, which
may fail if we find a device test manifest in a mixed module for which
a Facet is not the right unit of abstraction.

Bug: 147586218
Test: existing
Change-Id: I1da93e7ddf4ff9c64b09c706dfb7cdeaa01e12a6

GitOrigin-RevId: 36ba0163541edddc8d10718cfcc317c76c9e58ac