[android] IDEA-359225 change intellij.gradle.dsl.kotlin module package to avoid the classloading issue in intellij.android.gradle.dsl plugin
To follow the plugin V2 model structure, every module that is a content of the plugin has to specify the module package attribute in the module configuration file.

To specify the package attribute for 'intellij.android.gradle.dsl.kotlin', classes had to be moved to different package because of the classloader issue in 'gradle.dsl' plugin if the package was specified as a 'com.android.tools.idea.gradle.dsl.parser.kotlin'.

The reason for classloader issues is a package collision with classes in the 'intellij.android.gradle.dsl' module.
 The 'intellij.android.gradle.dsl' module contains the following files under the package 'com.android.tools.idea.gradle.dsl.parser.kotlin':
 - KotlinDslElement.java
 - CompilerOptionsDslElement.java
 and its module package attribute is set to 'com.android.tools.idea.gradle.dsl'.

 When the gradle.dsl plugin tries to load classes in the 'intellij.android.gradle.dsl' module's 'com.android.tools.idea.gradle.dsl.parser.kotlin' package, it requires 'intellij.android.gradle.dsl.kotlin' module's classloader to do so, because of the package matches its module package attribute.

 To avoid the classloader issue, 'intellij.android.gradle.dsl.kotlin' module classes are moved under different unique packages.

GitOrigin-RevId: 7a6aa7a233cd771ce6380b20e2d7cdbf53d39bd7