Add interfaces to support artifact metadata.
ArtifactMetadata represents a type of metadata that can be extracted
from a build artifact. ArtifactMetadataProvider allows the artifact
tracker to query which metadata is needed per build artifact.

ProjectProtoTransform implements ArtifactMetadataProvider to allow a
transform to specify which metadata is needed.

While we're here, also move ProjectProtoTransform into the top level
package to reflect the fact that it's scope is now wider than just the
project package.

Follow-up CLs will add logic to the artifact tracker to extract and
store this metadata, which will then allow project proto transforms
to request is later on.

Bug: 357876275
Test: No functional changes in this CL so no new tests yet.
Change-Id: I7a483ee91f8312bfc7a2cff606f395524f4eea71

GitOrigin-RevId: 5ce41a1654636415cce848f2edc2cb791a1a02db