Isolate build-system-specific parts of databinding support
The questions that the databinding layout would like to have answered
regard the test-like nature of a module (in order to find the right
package name), and whether any additional modules should be
considered (in order to find the right scope).  As usual, there would
be an argument that these are not the right questions, as they operate
on the wrong level of granularity in general, but we should support
them in the medium term.

Define a new module for the implementation of gradle-specific
databinding functionality.  (Add project system teams to OWNERS for
that module).  Define a token for the questions, and implement a
behaviour-preserving GradleToken to give non-default answers for
Gradle projects.

Bug: 147586218
Test: existing
Change-Id: I185c0139c05ca02346d91eb8bc70bcd19dd7a07e

GitOrigin-RevId: 6d5cf3518e498bab9e10e92e8037cdd7b75ce35f