NPE Protection For Root Node
When we click on a node, we first dispose the xml editor and
then try to get a new editor for the selected node.

If the selected node is the root ("/"),
`path.getData().getPath().getFileName()` returns null and we throw.

The green screen is caused by the existing editor that has been disposed
but is still displayed in its disposed state.

Fixes: 369315719
Test: Manually
Change-Id: Ie320fb86adc152c60612625ce6a2e69ae91c03f1

GitOrigin-RevId: 189e7ff550714ae4aa00dd0b3a7efb67c82d0e00