Remove redundancy from StudioAssetFileOpener
ResourceContent.getSampleDataUserDir relies on the module system's
getSampleDataDirectory() method.  For linked Gradle modules -- the only
modules where there might be a distinction bewteen a given module and
its "holder module" -- the relevant implementation is in
MainContentRootSampleDataDirectoryProvider, which is invariant to
which of a linked group of modules is given (because it itself finds
the holder module of its given module).  Therefore, the possible query
of the holderFacet in StudioAssetFileOpener.getDirectories() will
never provide any additional information.

Bug: 147586218
Test: existing
Change-Id: I6ce868431d39f50ca626e1b42a4621abf50821a3

GitOrigin-RevId: 99878f580ac812e538fd7235a430aaa9b13a106d