Restore Page Size and Live Update State
When reconnecting to a database inspector, we restore tabs that were open in the previous session.

We currently loos the state of the UI when we restore the tabs.
Specifically, the max-page combobox and the live-update checkbox.

This cl adds these values to the `TabDescription` classes and handles
the propagation of the saved values to the controllers and views
invloved.

There are 2 distinct sets of controller-view stacks:
1. For tables: TableController / TableViewImpl
2. For queries: SqliteEvaluatorController containing an inner TableController / TableViewImpl  stack

Fixes: 350526050
Test: Manually
Change-Id: Icd79e46798a1135343551e5e4cb77600ae8376ea

GitOrigin-RevId: df9b6e92d9b4bf0ea7416e219112f995339505ef