Refactor transformAndShowDiff to allow different ways of
... handling generated code. In order to comply with the new requirement
of generating previews without a diff view for some actions, we first
need to parameterize the callback of transformAndShowDiff, so it doesn't
necessarily show the diff view (although this is kept as the default
behavior).

This change also renames this function to generateCodeAndExecuteCallback
so it's more aligned with its new behavior.

Bug: 365736160
Test: N/A
Change-Id: I9e613dd2f7fb7ef5a0d785ea1c3232a175c57478

GitOrigin-RevId: 400b4f2015d496a7b02ad7024db12178e68db031