Create PsiDocumentationTargetProvider for SDK docs
Our current handling of fetching external framework documentation for
SDK classes is using a deprecated extension point, and we aren't able to
customize it as we need to. (We want customization so we can fetch
tailored docs from developer.android.com, rather than downloading a ton
of data and postprocessing in a brittle manner in Studio code.)

This change moves handling to the new supported extension points, aiming
to maintain parity between the two implementations.

Bug: n/a
Test: AndroidSdkDocumentationTargetProviderTest
Change-Id: I60083d433bfbf18e6b9d9744b14b1237b5a1c946

GitOrigin-RevId: e881e91d2a914c155c8411f333872674d396dd84