[SDK Index] Check disposed project before notification
We currently show a notification after sync if the project has a
blocking issue. To generate this notification we wait for the SDK Index
to be ready and once it is the list of dependencies in the project is
queried. It is possible that users close the project while waiting for
SDK Index to be ready and thus the project could be disposed in the
meantime.

This change adds a couple of checks to confirm the project has not been
disposed before continuing.

Bug: 360162088
Test: SdkIndexIssuesSyncListenerTest
Change-Id: Ic586529fa083a8ad8683f89a81aba222b1f2aa4b

GitOrigin-RevId: 3b7cdd98bc91cd3c286a5901ffe7e695890bde90