Remove obsolete logic in AndroidDocumentationProvider
This logic was added in 2013 with the goal of improving quick docs for
framework resources. It did this by (1) rendering custom information
about the resource values, and (2) appending "regular" javadocs from
developer.android.com.

The logic was then broken in 2014 by a change coming from an IntelliJ
merge, and has been broken since. The status quo is that this logic
might run and take time/resources, but has no effect on the rendered
quick documentation.

Re-enabling this logic as initially envisioned now no longer works well,
as the rendered result doesn't look very good after merging the two
sources. I'm opting to remove this logic now to get rid of the
complexity, and we'll consider the framework resource case moving
forward as we look at overall quick documentation improvements.

Bug: n/a
Test: manually verified, and quick docs have existing coverage
Change-Id: I9e378d071fbbd082c002ac2d09bdea11dee5e5cd

GitOrigin-RevId: b989db3c1ea93c228694a2ae24109a2d52f7a689