Stop using an unnecessary callback when rendering previews
The helper render method was a suspendable function, that waited
for the render to finish. So, its callback is not necessary and is
being replaced by simply executing its code after the await returns.

Also, the helper method itself stops being useful without having the
callback, so it's being removed and replaced by directly calling the
corresponding SceneManager's render method, and awaiting on it.

Bug: 361786751
Test: existing should pass, behaviour shouldn't change.
Change-Id: Ie6868d81133f8838f7b1293bae1169c250d2c6ec

GitOrigin-RevId: c231c9be48c776fa397ad3827519af3fa8525358