Introduce a WHSPanelController class that will
... be responsible for creating and showing the WHS Panel and handling
user notifications when the panel is not showing.

This CL is just a first step. The WHS Panel is created once per emulator
controller. In the next CLs it will be created each time it is shown.

This CL also removes WearHealthServicesPanelManager which used to create
and cache WHSPanels. This was done as the WHSPanel contained logic
around notifying the user of successful/failed actions when the panel
was no longer showing. This responsibility will be shifted to the
Controller instead.

Bug: 369767252
Test: //tools/adt/idea/wear-whs:intellij.android.wear-whs_tests
Change-Id: I2720ff13ef4b824db9c38d9e483480a55fbd08bc

GitOrigin-RevId: a7d339a8f64895bc1bfade0301cd7c2cdd8aa851