Limit the displayed number of issues in PSD
When the PSD has too many issues (thousands) it takes considerable
amount of time (tens of seconds) to display them, causing UI freezes.

This change limits the number of issues that are displayed in the PSD by
taken the first 300 based on severity priority, path and text (in that
order). A warning is logged when there are more issues but were not
shown.

Bug: 344958680
Test: Manually checked that the UI is more responsibe and that the log
is created.

Change-Id: I818a55bd9a49e78bf3d34535be3a46b15f6e298a

GitOrigin-RevId: bd74dc8cecd5a04add61528830a21f8af48794c6