[SDK Index] Add update fixes to PSD issues
Limiting the number of issues in PSD is causing that update fixes are
not shown since they are presented as info issues. When there are
multiple errors and warnings, those issues get precedence and cuase the
quick fixes shown in the updates to not being visible.

This change adds update quickfixes directly into the SDK Index issues,
making them available even if the update issues are not shown. This also
has the advantage to have those fixes right were the error message is,
saving users looking in the list for them.

There are also some presentation changes:

- Add line separators when converting issue text to HTML (b/369997141)
- Move quick fixes into a new line
- Allow quick fix actions to have multiple lines (to be able to show
  quick fixes like "Update variable to 10.4.1" on a narrow button)

Bug: 344958680
Fixes: 369997141
Test: SdkIndexPsd*Test, PsAnalyzerDaemonKtTest
Test: DependencyViewIssueRendererTest, manually checked UI and freeze time

Change-Id: I66f2f3d39513f88a48ded4b19317d753eaa59ff7

GitOrigin-RevId: 392276cc8f23ef2216da930bd5f1c47dd7a889a2