Fix invariant in BuildIssueComposer
Before this change there was inconsistency: we were adding line return
after details line but not after the quickfix line. This leads to
confusion of the state of crafted message and unexpected empty lines
in the result.
Change to never add line return after an added piece. Add method for
inserting empty line explicitly when necessary.

Bug: N/A
Test: Existing Build output tests.
Change-Id: I234622f5a1d4d9446a61c9ac2742c09c268eaa30

GitOrigin-RevId: 64a39d9f6d64200738fab2d44d59195e9d01ce40