Send instruction as System Message
When building a prompt for SendPreviewToStudioBotAction, we first send a
message to the model to make it aware that it should act as an expert
Android programmer. This was planned to be a system message but, because
of a limitation in the previous model used, we had to inline it in the
user message. Since the model was updated, we can rollback the original
design.

Bug: N/A
Test: N/A. Manual.
Change-Id: Ib9771aec96b3fce66ac45ef481c79ac696204267

GitOrigin-RevId: 60515ced0d5592fb7b7b4a75e0e515a4e584a79f