Fix deprecation completion test for K2.
This test ensures that functions that are deprecated starting with
a specified API level are, in the completion dialog, only displayed as
deprecated (crossed out) when the API level of the surrounding code is
at or above the API level at which the function was deprecated. This was
accomplished by checking android.text.Html.fromHtml, which had a new
version introduced, and its previous version deprecated, in API 24.

However, choosing that function also inadvertently made this a test of
a Kotlin language feature called "SAM adapters". These are synthetic
functions, generated by the Kotlin compiler, when calling a Java
function that takes an interface with a single abstract method. Since
these SAM interfaces commonly represent a function in Java (which has no
native function type), the compiler allows calling the function with
a Kotlin lambda, and emits a conversion to the correct SAM interface
type under the hood.

In the K1 plugin, these SAM adapter methods appeared in completion,
meaning that any Java function that took SAM interfaces would appear
twice - once with the SAM type, and once with a Kotlin functional type.
Since Html.fromHtml has SAM-typed parameters (imageGetter and
tagHandler), this means that the completion test needed to account for
both the original method and the SAM adapter method.

However, in K2, these SAM adapters no longer appear in completion. It's
unclear whether this is a deliberate change, or missing functionality in
the K2 IJ plugin. This causes the test to fail, because the expected SAM
adapter method is now missing in the completion suggestions.

To sidestep this entire issue and make the test pass on both K1 and K2,
this change replaces the API being tested with another API with similar
characteristics (a new overload introduced in a given API, deprecating
the original overload), but without SAM-typed parameters.

Since this is the last failing K2 test in the module, this change also
enables K2 testing in presubmit.

Fixes: 353683353
Test: bazel test //tools/adt/idea/android/jetbrains:all
Change-Id: Ia053f6bf97eb3280f8d9b71262ad50208a066ecb

GitOrigin-RevId: 9579b1c08f0f768cca62a563318c74f767ac6089