Make Compose state node expandable only when needed
This CL tweaks the ComposeStateNode presentation; instead of always
reporting that is has children (which make it an expandable tree node in
the debugger UI), it only does that when the childNodes list is not
empty. This avoids having an expandable node with no children, which
looks a bit broken.

This would mostly happen in simple Compose UIs, or during the first
composition pass, when usually no state info is attached.

Bug: none
Test: no unit tests, smoke tested (difficult to unit test)
Change-Id: I24dd068f8e0196824412a7da266367c5293845dd

GitOrigin-RevId: e304c6a274ff1f525fee233a79f5e3fa2df7e8a3