Prepare for move of getMainModule to Gradle project system
There is still a need to be able to convert between holder modules,
returned from things like getAndroidFacets(), and production (Android)
source sets.  Define getProductionAndroidModule() as a module system
method, with a specialized implementation in the Gradle module system.
Make it nullable, because in principle we might be working with a
non-Android entity.

We sometimes know that we *are* in fact dealing with an Android
entity, usually when we have an AndroidFacet; define a utility
non-nullable function for those circumstances.

Adjust spellings of calls to getMainModule in non-Gradle-related
modules to use the appropriate one of these two new methods.

Bug: 147586218
Test: existing
Change-Id: I054e9a7881fa0fcd2b22c86875ff5e224d04b3f3

GitOrigin-RevId: 0b6c4904d64450072d64355c796f6f4a189c8627