Add fast path in SdkWritingAccessProvider
Querying isInAndroidSdk() can be slow (~few hundred milliseconds) the
first time it is called because RootIndex.buildRootInfo() needs to
load the dependency graph. This can happen at inconvenient times, such
as when the user is trying to type something in the editor.

We can avoid this latency in the common case by checking isInContent()
before querying isInAndroidSdk(). The former tends to be faster, and
is already called on the EDT by IntelliJ services such as
NonProjectFileWritingAccessProvider.

Bug: 371950559
Test: SdkWritingAccessProviderTest
Change-Id: I2f71797fb6efc1f03474ab99ccb925de3841be77

GitOrigin-RevId: e5bac3be4d8eb1114b6e2c7180568d729882bbf1