Fix propagation of system images
The set of system images that we display in the table is dependent on
the set of images we have, and on the state of the filters. Both of
these can change; previously the DevicePanelState assumed that the
system images were static.

Instead of storing the images as mutable state in DevicePanelState, pass
the immutable SystemImageState from the top level down to where they are
needed, and apply the filters when necessary.

Fixes: 372074906
Fixes: 371272376
Test: LocalVirtualDeviceSourceTest
Change-Id: Ie9f4dbba63ba03a94867905950784826171c6c59

GitOrigin-RevId: 6eda79d7e66007ef674fb53842820e2491b27e52