Fix flaky tests
The tests were failing during the setup due to an
UnfinishedStubbingException when mocking the action manager.

I'm not too sure why that was happening only on windows. I think it
might have been caused by some service accessing the action manager
before it was finished being mocked.

This CL completes the mocking of the action manager before substituting
the real action manager service.

Fixes: 372020257
Test: //tools/adt/idea/wear-whs:intellij.android.wear-whs_tests * 1000
Change-Id: I20a580e23be75c6db56fd7d68fe023b44d7da0f2

GitOrigin-RevId: 17dc25ad4ced70ce706b9a9af5570c1db82cba81