Workaround bazel duplicate depset entries issue.
There are cases when bazel can yield apparently duplicate File entries
in a depset. This can occur when two different aspects generate the same
file. This then causes problems with duplicate entries in the targetInfo
map in the artifact tracker, and two `TargetBuildInfo` entries are not
considered identical if one has duplicate file entries.

Fix this issue by storing the `BuildArtifact`s in a set rather than a
list.

Test: Tested manually.
Test: Automated test blocked by b/338177779
Bug: 371973613

Change-Id: If7f5bffbb9e6808863b3491a0659e4513c1fd8b9

GitOrigin-RevId: 1d008fa4e45dfcf056e1e83de52a5a6d40f1a55f