[PSD] Stronger disposable handling
We have reports of mySelectedConfigurable retaining a disposed
configurable.  It's hard for me to see how this happens, but: let us
defend against it by registering a child disposable to clear the
field.

If that can happen, then presumably so can holding on to a disposed
configurable from myConfigurables.  Again, register a child disposable
to clear a key from the map on its disposal.

MyDisposable is also lazily updated after its disposal.  Update it
eagerly instead.

Bug: 372198739
Test: maybe we'll get fewer crash reports
Change-Id: If74d7a38feab3381548ea8989411e7e47b139a76

GitOrigin-RevId: ec8badfdf66d2fa46c49e8e292cd32fbb01364c2