Add space at the bottom of GridLayoutManager
This change is a copy from change 2ac18c6 with conflicts fixed and with  the framePadding replaced with the bottomPadding.
If horizontal scrollable is available, preview would overlap, a bit of
extra canvas padding at the bottom make the most bottom preview not to
be that close to the bottom edge.

Test: visually verified
Bug: n/a
Change-Id: Ibdc41334524b8a7e18d2ebc9e995d35241d42deb

GitOrigin-RevId: 2394be99a04154dc567b488dc48c464ee296b45d