[Dsl] Specialize support for Gradle properties
As distinct from language properties: specifically Property<T>
entities.  We already have reader support for `.set(...)`, so what
remains is to create new elements in that form, rather than through
assignment.  This involves writing (twice, because we can punt for
Declarative Gradle for now) the slightly awful Psi manipulations that
arrange that create / apply have the same understanding of the
incomplete Psi that results.

Use all this to support repositoriesMode, since Gradle before 8.2
doesn't appear to support the assignment syntax.

Bug: 368629093
Test: existing test adjusted
Change-Id: I19c505483e618800aa20574b61524d3b1fca3a80

GitOrigin-RevId: 9869c89f89a0aa4beabf40eced774352558841fd